[SelectMenu] Add basic support for form participation.
This CL adds basic form participation support by having
HTMLSelectMenuElement extend HTMLFormControlElementWithState that already
providies many behaviors (e.g. form association).

For now similar to <select>, name attribute is going to be required
for form participation and the value is going to be provided by
selectmenu.value.

Support for saving and restoring form control state and validation related
is going to be added in future CLs.

Bug: 1121840
Change-Id: Id9846a1fe9b781db5e4011fec77ff47faec2aaad