Fix element hash fallback in iframes
The text directive element-id fallback tries to scroll to the element-id
if a text match wasn't found. It does so in PerformPreRafActions since
element fragments cause focus which can execute script. Invoke(), where
the text fragment performs its text search, occurs from lifecycle
updates where script execution is forbidden.

Text directives used to be limited to main-frames but were enabled
explicitly for iframes in https://crrev.com/a71830a90b72f0a613fe9.
However, PerformPreRafActions wasn't updated to call into the
FragmentAnchor of an iframe

We also need to strip the fragment directives in
ElementFragmentAnchor::TryCreate.  In cases where a text fragment is
successfully created, the element id fallback works by creating an
ElementFragmentAnchor from inside TextFragmentAnchor. However, if the
TextFragmentAnchor is blocked or not parsed correctly, the URL passed to
ElementFragmentAnchor in FragmentDirective::TryCreate will still include
the directives. A test for this was added in the new
same-document.tests.html

Bug: 1260135
Change-Id: Ib3f19b462016d34cc41d21b05933aa9cb822a9b8