Find OOFs in the fragment tree when looking for a scroll anchor.
The HasOutOfFlowFragmentChild() bits are taken from
https://chromium-review.googlesource.com/c/chromium/src/+/2640403 by
obrufau@igalia.com, and the changes in ScrollAnchor is a counterproposal
to how that CL solves it.

In LayoutNG block fragmentation, OOF fragments are normally placed as
direct children of the fragmentainers, so that's where we need to look
for them.

Additionally, in cases where we're not inside block fragmentation, look
for OOF child fragments inside the fragment(s) generated by the
candidate. This is in preparation for the CL mentioned above (which will
no longer populate LayoutBlock's PositionedObjects when LayoutNG is
used).

OOFs are easily missed in ScrollAnchor::Examine(), because OOFs don't
contribute to layout overflow, and this is why we needed this extra
effort in the first place (crbug.com/692701).

This fixes wpt/css/css-scroll-anchoring/abspos-in-multicol-001.html
when LayoutNGBlockFragmentation is enabled.

Bug: 829028
Test: external/wpt/css/css-scroll-anchoring/abspos-in-multicol-001.html
Change-Id: I25bd7b3a56b2405ec76823993eefed4889361250