Add web platform tests for HTTP partitioned cookies
The test is expected to fail in the standard test suite and pass in the
virtual test suite when partitioned cookies are enabled.

This test uses the Partitioned attribute to indicate a cookie is
partitioned, but it would also pass if cookies are partitioned by
default as well.

We can also use the virtual test suite I added here to test the changes
to the CookieStore API in https://crrev.com/c/3212611.

Bug:1225444
Change-Id: I3f0c0907da8febee559b2211b00420e8767a48fe