App history: make it mostly nonfunctional for opaque origins
Given that opaque origins complicate the notion of what the origin of a
history entry is and when they can be considered contiguous, discussion
on https://github.com/WICG/app-history/issues/167 suggests that we
should probably start with restricting the API to not work on them for
now, in the same way we do for the initial about:blank.

This also slightly improves the test coverage for detached window
scenarios since they are very similar to what we're changing here.

Fixed: 1252490
Change-Id: Id3c761d87dba0c3b75c0e70e4f6cc61e96429b8d